jimi::crc32c_hw_x64()返回类型修改为uint32_t;

+++ src/main/crc32/crc32c.h
@@ -72,7 +72,7 @@ static uint32_t crc32c_hw_x86(const char * data, size_t length)
 
 #if CRC32_IS_X86_64
 
-static uint64_t crc32c_hw_x64(const char * data, size_t length)
+static uint32_t crc32c_hw_x64(const char * data, size_t length)
 {
     assert(data != nullptr);
 
@@ -103,7 +103,7 @@ static uint64_t crc32c_hw_x64(const char * data, size_t length)
         }
     } while (1);
 
-    return ~crc64;
+    return (uint32_t)~crc64;
 }
 
 #endif // CRC32_IS_X86_64
